backend/pyproject.toml
[project]
name = "tsosi-backend"
version = "0.1.0"
requires-python = ">=3.12"
description = "TSOSI application backend code."
readme = "README.md"

[tool.poetry]
package-mode = false
homepage = "https://tsosi.org/"
repository = "https://github.com/tsosi-org/tsosi-app/"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
# HTTP server
gunicorn = "^23.0.0"
# API & DB deps
Django = "~5.1.2"
djangorestframework = "~3.15.2"
django-filter = "~24.3"
django-cors-headers = "~4.6.0"
psycopg = "~3.2.3"
# Backend jobs deps
pandas = "~2.2.3"
aiohttp = "~3.10.9"
openpyxl = "~3.1.5" # Pandas requires this package but it's not in the dependencies 
pycountry = "~24.6.1"
requests = "~2.32"
pillow = "~11.0.0"
redis = "^5.2.1"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.7.0"

[tool.poetry.group.dev.dependencies]
# Dev linting & formatting
ipython = "^8.28"
ipykernel = "^6.29"
black = "^24.10.0"
isort = "^5.13.2"
# Tests
pytest = "^8.3.4"
pytest-django = "^4.9.0"
factory-boy = "^3.3.3"
# Deployment utilities
scp = "^0.15.0"
paramiko = "^3.5.0"
# Random stuff
termcolor = "^2.5.0"
xlsxwriter = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend_site.settings"
# django_find_project = false
python_files = [
  "*_test.py",
  "*_tests.py",
  "test_*.py",
  "tests_*.py"
]